Update README.markdown to show default search path
When mounting the soulmate engine into a Rails app, at '/sm', the README
indicated that this is the URL we should query.  This is incorrect for the gem
itself, which expects you to query '/sm/search'.  The demo app works because they
have obviously mounted the engine differently (w/o 'search'). 